[+] move comments to users/initiatives
